+++ python-lib/sharepoint_fresh_token.py
@@ -1,0 +1,77 @@
+from safe_logger import SafeLogger
+from dss_constants import DSSConstants
+import time
+import threading
+
+logger = SafeLogger("sharepoint-online plugin FreshToken", DSSConstants.SECRET_PARAMETERS_KEYS)
+TOKEN_VALIDITY_SAFETY_MARGIN_SECONDS = 60
+
+# Must have the same properties than dataiku.core.plugin.OAuthCredentials (access_token)
+class FreshToken():
+    def __init__(self, token_refresh_method=None, access_token=None):
+        logger.info("FreshToken init")
+        if not access_token and not token_refresh_method:
+            logger.error("No valid token or token refresh method were provided")
+            raise Exception("No valid token or token refresh method were provided")
+        if access_token:
+            logger.info("Permanent access token provided")
+            self.current_token = access_token
+            self.token_refresh_method = self._default_refresh_method
+            self.token_renewal_time = None
+        self._refresh_lock = threading.Lock()
+        if token_refresh_method is not None:
+            logger.info("Using refresh method")
+            self.token_refresh_method = token_refresh_method
+            self.refresh_token()
+
+    def _default_refresh_method(self):
+        return self.current_token
+
+    def token_needs_renewal(self):
+        if self.token_renewal_time is None:
+            return False
+        epoch_time_now = int(time.time())
+        return self.token_renewal_time <= epoch_time_now
+
+    def refresh_token(self):
+        self.current_token = self.token_refresh_method()
+        if not self.current_token:
+            raise Exception("The access token could not be refreshed")
+        self.token_renewal_time = get_token_renewal_time(self.current_token)
+        logger.info("The token is valid until {}".format(self.token_renewal_time))
+
+    @property
+    def access_token(self):
+        if not self.token_needs_renewal():
+            return self.current_token
+        with self._refresh_lock:
+            if self.token_needs_renewal():
+                logger.info("Token reaching its time limit, refreshing it...")
+                self.refresh_token()
+        return self.current_token
+
+
+def get_token_renewal_time(token):
+    decoded_jwt = decode_jwt(token)
+    token_renewal_time = decoded_jwt.get("exp", None)
+    if isinstance(token_renewal_time, int):
+        token_renewal_time = token_renewal_time - TOKEN_VALIDITY_SAFETY_MARGIN_SECONDS
+    return token_renewal_time
+
+
+def decode_jwt(jwt_token):
+    try:
+        import base64
+        import json
+        sub_tokens = jwt_token.split('.')
+        if len(sub_tokens) < 2:
+            logger.error("JWT format is wrong")
+            return {}
+        token_payload = sub_tokens[1]
+        padded_token = token_payload + "=" * (-len(token_payload) % 4)
+        decoded_token = base64.urlsafe_b64decode(padded_token.encode('utf-8'))
+        json_token = json.loads(decoded_token)
+        return json_token
+    except Exception as error:
+        logger.error("Could not decode JWT token ({})".format(error))
+    return {}